+++ packages/datadog-plugin-aws-sdk/test/spec_helpers.js
@@ -15,6 +15,24 @@ const helpers = {
       delete process.env['AWS_SECRET_ACCESS_KEY']
       delete process.env['AWS_ACCESS_KEY_ID']
     })
+  },
+
+  dsmStatsExist (expectedHash) {
+    const dsmStats = agent.getDsmStats()
+    let hashFound = false
+    if (dsmStats.length !== 0) {
+      dsmStats.forEach((statsTimeBucket) => {
+        statsTimeBucket.Stats.forEach((statsBucket) => {
+          statsBucket.Stats.forEach((stats) => {
+            if (stats.Hash.toString() === expectedHash.readBigUInt64BE(0).toString()) {
+              hashFound = true
+              return hashFound
+            }
+          })
+        })
+      })
+    }
+    return hashFound
   }
 }
